+++ src/behaviors/index.js
@@ -1,10 +1,8 @@
 /**
  * Point of contact for behaviors.
  * @module behaviors
- * @example import { Document } from './models';
+ * @example import behaviors from './behaviors';
  */
-
-import { assign, pick, values } from 'lodash';
 
 import { id_from_title } from './id_from_title/id_from_title';
 
@@ -15,13 +13,4 @@ const behaviors = {
   ...('behaviors' in config ? config.behaviors : {}),
 };
 
-/**
- * Apply behaviors to document.
- * @method applyBehaviors
- * @param {Object} document Document model.
- * @param {Array} newBehaviors Array of behaviors to be applied.
- * @returns {Object} Document with behaviors applied.
- */
-export function applyBehaviors(document, newBehaviors) {
-  return assign(document, ...values(pick(behaviors, newBehaviors)));
-}
+export default behaviors;