+++ CHANGELOG.md
@@ -38,6 +38,7 @@
 - Upgraded dependencies @robgietema
 - Update node dependency requirement @robgietema
 - Split profiles in core and default @robgietema
+- Move apply behaviors to the document class @robgietema
 
 ### Documentation